mengubahproduk,pengeluaranlain,detailpemesanan,pembelian,database

+++ database/migrations/2024_05_04_003642_detail_pemesanan.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id('id');
             $table->foreignId('id_pemesanan')->constrained('pemesanans');
             $table->foreignId('id_produk')->constrained('produks');
-            $table->integer('jumlah');
+            $table->double('jumlah');
             $table->timestamps();
         });
     }
